+++ tests/test_ticker.py
@@ -3,9 +3,7 @@ import itertools
 import os
 from datetime import datetime
 
-from yahooquery import Research, Screener, Ticker
-from yahooquery import (
-    get_currencies, get_exchanges, get_market_summary, get_trending)
+from yahooquery import Ticker
 
 
 TICKERS = [
@@ -13,16 +11,13 @@ TICKERS = [
            password=os.getenv('YF_PASSWORD')),
     Ticker('aapl ^GSPC btcusd=x brk-b logo.is', asynchronous=True),
     Ticker(['aapl', 'aaapl']), Ticker('hasgx'),
-    Ticker('btcusd=x', formatted=True)
-]
-
-RESEARCH = [
-    Research(
-        username=os.getenv('YF_USERNAME'), password=os.getenv('YF_PASSWORD'))
+    Ticker('btcusd=x', formatted=True, validate=True)
 ]
 
 FINANCIALS = ['cash_flow', 'income_statement', 'balance_sheet',
-              'p_cash_flow', 'p_income_statement', 'p_balance_sheet']
+              'p_cash_flow', 'p_income_statement', 'p_balance_sheet',
+              'all_financial_data', 'p_all_financial_data',
+              'p_valuation_measures']
 
 SEPERATE_ENDPOINTS = FINANCIALS + [
     'option_chain', 'history', 'all_modules',
@@ -45,11 +40,6 @@ def test_premium(ticker, prop):
 
 @pytest.fixture(params=TICKERS)
 def ticker(request):
-    return request.param
-
-
-@pytest.fixture(params=RESEARCH)
-def research(request):
     return request.param
 
 
@@ -83,6 +73,10 @@ def test_multiple_modules_str(ticker):
     assert ticker.get_modules("assetProfile summaryProfile") is not None
 
 
+def test_news(ticker):
+    assert ticker.news() is not None
+
+
 def test_all_modules(ticker):
     assert ticker.all_modules is not None
     data = ticker.all_modules
@@ -98,6 +92,33 @@ def test_modules(ticker, module):
     el for el in itertools.product(FINANCIALS, ['q', 'a'])])
 def test_financials(ticker, frequency, module):
     assert getattr(ticker, module)(frequency) is not None
+
+
+def test_bad_financials_arg():
+    with pytest.raises(KeyError):
+        assert Ticker('aapl').income_statement('r')
+
+
+def test_no_financials_data(ticker):
+    assert 'data unavailable' in Ticker('BC94.L').income_statement()
+
+
+def test_get_financial_data(ticker):
+    assert ticker.get_financial_data([
+        'GrossProfit',
+        'NetIncome',
+        'TotalAssets',
+        'ForwardPeRatio'
+    ]) is not None
+
+
+def test_p_get_financial_data(ticker):
+    assert ticker.p_get_financial_data([
+        'GrossProfit',
+        'NetIncome',
+        'TotalAssets',
+        'ForwardPeRatio'
+    ]) is not None
 
 
 @pytest.mark.parametrize("period, interval", [
@@ -123,75 +144,3 @@ def test_history_start_end(ticker, start, end):
 def test_history_bad_args(ticker, period, interval):
     with pytest.raises(ValueError):
         assert ticker.history(period, interval)
-
-
-def test_get_currencies():
-    assert get_currencies() is not None
-
-
-def test_get_exchanges():
-    assert get_exchanges() is not None
-
-
-def test_get_market_summary():
-    assert get_market_summary() is not None
-
-
-def test_get_trending():
-    assert get_trending() is not None
-
-
-def test_screener():
-    s = Screener()
-    assert s.get_screeners('most_actives') is not None
-
-
-def test_available_screeners():
-    s = Screener()
-    assert s.available_screeners is not None
-
-
-def test_bad_screener():
-    with pytest.raises(ValueError):
-        s = Screener()
-        assert s.get_screeners('most_active')
-
-
-def test_reports(research):
-    assert research.reports() is not None
-
-
-def test_reports_one_arg(research):
-    assert research.reports(report_date='Last Week') is not None
-
-
-def test_reports_multiple_filters(research):
-    assert research.reports(
-        investment_rating='Bearish,Bullish',
-        report_date='Last Week',
-        sector=['Basic Materials', 'Real Estate'],
-        report_type='Analyst Report'
-    ) is not None
-
-
-def test_reports_bad_arg(research):
-    with pytest.raises(ValueError):
-        assert research.reports(investment_type='Bearish')
-
-
-def test_reports_bad_option(research):
-    with pytest.raises(ValueError):
-        assert research.reports(report_type='Bad Report Type')
-
-
-def test_reports_bad_multiple(research):
-    with pytest.raises(ValueError):
-        assert research.reports(report_date=['Last Week', 'Last Year'])
-
-
-def test_trades(research):
-    assert research.trades() is not None
-
-
-def test_trades_size(research):
-    assert research.trades(300) is not None